// src/app/model/info-gestio-reg.ts
export class InfoGestioReg {
    id:                 number;
    periode:            String;
    codiEmpresa:        String;
    tipusProducte:      String;
    dataUltimReg:       Date;
    pendent:            Boolean;
    registrat:          Boolean;
    noComercialitzacio: Boolean;
    tancat:             Boolean;
}